tests/wpt/web-platform-tests/html/semantics/embedded-content/the-img-element/delay-load-event.html

<html>
<head>
<title>Image element delays window's load event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<img src="resources/cat.jpg?pipe=trickle(d2)">
<script>
    async_test(function(t) {
        var saw_img_load = false;
        document.querySelector('img').onload = t.step_func(function() {
            saw_img_load = true;
        });
        addEventListener('load', t.step_func_done(function() {
            assert_true(saw_img_load);
        }));
    });
</script>
